vcs-log: migrate log tabs to the tool window
Rebased keeps Version Control in a dedicated central host, so VCS Log tabs must no longer compete with the File Editor host.

Remove the obsolete VCS Log virtual-file and editor-provider path, migrate persisted editor tabs to the tool window on load, and normalize SHOW_IN_EDITOR to false. Keep tab restoration, actions, settings, and empty-editor shortcuts aligned with the tool-window content manager.
