+++ platform/vcs-log/impl/resources/messages/VcsLogBundle.properties
@@ -9,7 +9,6 @@ action.Vcs.Log.ShowLongEdges.description=Show long branch edges even if commits 
 action.Vcs.Log.ShowOnlyAffectedChanges.description=Only show changes that affect files selected in the "Paths" menu
 action.Vcs.Log.ShowTagNames.description=Show tag names in the table
 vcs.log.action.description.open.new.tab.with.log=Open new tab with {0} Log
-vcs.log.action.description.open.new.tab.with.log.in.editor=Open new tab with {0} Log in editor
 action.Vcs.Log.ShowDiffPreview.text=Show Diff Preview
 action.Vcs.Log.MoveDiffPreviewToBottom.text=Bottom
 action.Vcs.Log.MoveDiffPreviewToBottom.description=Locate Diff Preview in the bottom
@@ -24,12 +23,10 @@ action.Vcs.Log.ShowOnlyAffectedChanges.text=Show Only Affected Changes
 action.Vcs.Log.ShowRootsColumnAction.text=Root Names
 action.Vcs.Log.ShowTagNames.text=Tag Names
 action.Vcs.Log.ShowIssuePreviewOnHover.description=Show the issue preview on hover over an issue id in
-action.Vcs.Log.ShowInEditor.description=Show the log in the editor window
 action.Vcs.Log.PreferCommitDate.text=Commit Timestamp
 prefer.commit.timestamp.action.text.show=Show Commit Timestamp
 action.Vcs.Log.PreferCommitDate.description=Show the time when the change was committed rather than authored
 vcs.log.action.open.new.tab.with.log=Open New {0} Log Tab
-vcs.log.action.open.new.tab.with.log.in.editor=Open New {0} Log Tab
 action.process.expanding.linear.branches=Expanding linear branches\u2026
 action.process.expanding.merges=Expanding merges\u2026
 action.description.expand.merges=Expand merges
@@ -131,7 +128,7 @@ vcs.log.graph.arrow.tooltip.jump.to.hash=Jump to commit {0}
 vcs.log.graph.arrow.tooltip.jump.to.hash.in.root=Jump to commit {0} in {1}
 vcs.log.references.more.tooltip=... {0} more in details pane
 
-# log in editor
+# shared virtual file type used by VCS-related previews
 filetype.vcs.log.description=VCS Log
 
 # replacement for specific vcs name
@@ -153,7 +150,6 @@ vcs.log.no.commits.matching.status=No commits matching filters
 vcs.log.reset.filters.status.action=Reset filters
 vcs.log.refresh.status.action=Refresh
 vcs.log.default.status=Changes log
-vcs.log.duplicated.tab.id.error=VCS Log tab cannot be opened twice
 
 # file history
 file.history.error.status=Error calculating file history
@@ -192,7 +188,6 @@ vcs.log.applying.filters.process=Applying Filters\u2026
 vcs.log.tab.name=Log
 vcs.log.diff.preview.editor.empty.tab.name=Repository Diff
 vcs.log.diff.preview.editor.tab.name=Repository Diff: {0}
-vcs.log.editor.name=VCS Log Editor
 
 # filters
 vcs.log.filter.all=All
